Converge towards i386.  I originally resisted creating <machine/pc/bios.h> because it was mostly irrelevant - except for the silly BIOS_PADDRTOVADDR etc macros.  Along the way of working around this, I missed a few things.
* Make syscons properly inherit the bios capslock/shiftlock/etc state like
  i386 does.  Note that we cannot inherit the bios key repeat rate because
  that requires a bios call (which is impossible for us).
* Give syscons the ability to beep on amd64.  Oops.

While here, make bios.c compile and add it to files.amd64.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@135690 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/amd64/amd64/bios.c
@@ -37,6 +37,9 @@ __FBSDID("$FreeBSD$");
 #include <sys/param.h>
 #include <sys/systm.h>
 #include <sys/kernel.h>
+#include <vm/vm.h>
+#include <vm/pmap.h>
+#include <machine/vmparam.h>
 #include <machine/pc/bios.h>
 
 #define BIOS_START	0xe0000
@@ -84,7 +87,7 @@ bios_sigsearch(u_int32_t start, u_char *sig, int siglen, int paralen, int sigofs
 	/* compare here */
 	if (!bcmp(sp + sigofs, sig, siglen)) {
 	    /* convert back to physical address */
-	    return((u_int32_t)BIOS_VADDRTOPADDR(sp));
+	    return((u_int32_t)(uintptr_t)BIOS_VADDRTOPADDR(sp));
 	}
 	sp += paralen;
     }
